refactor: rename .ts files into .mts
