+++ src/helpers/handleCdCommand.mts
