+++ src/helpers/handleLsCommand.mts
