+++ src/helpers/handleUpCommand.mts
