+++ tsconfig.json
@@ -9,7 +9,7 @@
         "skipLibCheck": true,
         "esModuleInterop": true,
         // "moduleResolution": "node",
-        "types": ["node"],
+        "types": ["node"]
     },
     "include": ["src/**/*"],
     "exclude": ["node_modules"]
